+++ metadata/tracking/listenbrainz/listenbrainz_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 
 	"github.com/golusoris/goenvoy/metadata"
-	"github.com/golusoris/goenvoy/metadata/music/listenbrainz"
+	"github.com/golusoris/goenvoy/metadata/tracking/listenbrainz"
 )
 
 func setup(t *testing.T, handler http.HandlerFunc) *listenbrainz.Client {